Extract AsyncMap to separate class
